app/requirements.txt
requests~=2.27.1
Pillow~=9.1.0
paho-mqtt~=1.6.1
Flask~=2.1.1
flasK_sqlalchemy~=2.5.1
python-dotenv~=0.15.0
flask-login~=0.6.0
IPy
svgwrite
python-dateutil
pytz